速度800总体还行，断路十字可能出错，右环岛出环岛可能出赛道

+++ project/code/judgement.h
@@ -29,7 +29,7 @@ extern RaceStatus CURRENT_STATUS;
 extern uint8 left_circle_flag, right_circle_flag, cross_flag;
 extern uint8 circle_threshold, cross_threshold;
 extern uint8 circle_status;
-extern uint8 cross_cnt;
+extern uint16 cross_cnt;
 extern bool out_flag;
 
 #endif /* JUDGEMENT_H_ */
